+++ Entities/ACDropShip.h
@@ -342,8 +342,6 @@ protected:
 
     // Automatically stabilize the craft with the upper thrusters? Defaults to yes.
     int m_AutoStabilize;
-    // The craft explodes if it has been on its side for more than this many MS (default 4000). Disable by setting to -1.
-    float m_ScuttleIfFlippedTime;
 
 	// Maximum engine rotation in degrees
 	float m_MaxEngineAngle;