+++ Entities/ACRocket.h
@@ -321,8 +321,6 @@ protected:
     // Limb paths for different movement states.
     // [0] is for the right limbs, and [1] is for left.
     LimbPath m_Paths[2][GearStateCount];
-    // The craft explodes if it has been on its side for more than this many MS (default 4000). Disable by setting to -1.
-    float m_ScuttleIfFlippedTime;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////